修改chainid支持范围

+++ src/main/java/io/hpb/web3/crypto/TransactionDecoder.java
@@ -20,7 +20,7 @@ public class TransactionDecoder {
         BigInteger value = ((RlpString) values.getValues().get(4)).asBigInteger();
         String data = ((RlpString) values.getValues().get(5)).asString();
         if (values.getValues().size() > 6) {
-            byte v = ((RlpString) values.getValues().get(6)).getBytes()[0];
+            int v = ((RlpString) values.getValues().get(6)).getBytes()[0];
             byte[] r = Numeric.toBytesPadded(
                 Numeric.toBigInt(((RlpString) values.getValues().get(7)).getBytes()), 32);
             byte[] s = Numeric.toBytesPadded(
